# API_DOCUMENTATION.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Finance Data Processing API",
                 "description":  "Backend API for Finance Dashboard with Role-Based Access Control",
                 "contact":  {
                                 "name":  "Md Shahbaz Imam",
                                 "email":  "mdshahbazimam02@gmail.com"
                             },
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Default Server"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/dashboard/trends":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "Dashboard"
                                                                     ],
                                                            "summary":  "Get monthly trends",
                                                            "description":  "Returns monthly income and expense breakdown. Default is last 6 months. ANALYST and ADMIN only.",
                                                            "operationId":  "getMonthlyTrends",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "months",
                                                                                   "in":  "query",
                                                                                   "required":  false,
                                                                                   "schema":  {
                                                                                                  "maximum":  24,
                                                                                                  "minimum":  1,
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int32",
                                                                                                  "default":  6
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/ApiResponseListMonthlyTrendResponse"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      }
                                                                          },
                                                            "security":  [
                                                                             {
                                                                                 "bearerAuth":  [

                                                                                                ]
                                                                             }
                                                                         ]
                                                        }
                                            },
                  "/api/dashboard/summary":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "Dashboard"
                                                                      ],
                                                             "summary":  "Get overall financial summary",
                                                             "description":  "Returns total income, total expenses, net balance, and transaction count. Accessible by all roles.",
                                                             "operationId":  "getSummary",
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponseDashboardSummaryResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           },
                                                             "security":  [
                                                                              {
                                                                                  "bearerAuth":  [

                                                                                                 ]
                                                                              }
                                                                          ]
                                                         }
                                             },
                  "/api/dashboard/recent":  {
                                                "get":  {
                                                            "tags":  [
                                                                         "Dashboard"
                                                                     ],
                                                            "summary":  "Get recent transactions",
                                                            "description":  "Returns the latest N transactions. Default limit is 10. Accessible by all roles.",
                                                            "operationId":  "getRecent",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "limit",
                                                                                   "in":  "query",
                                                                                   "required":  false,
                                                                                   "schema":  {
                                                                                                  "maximum":  100,
                                                                                                  "minimum":  1,
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int32",
                                                                                                  "default":  10
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/ApiResponseListTransactionResponse"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      }
                                                                          },
                                                            "security":  [
                                                                             {
                                                                                 "bearerAuth":  [

                                                                                                ]
                                                                             }
                                                                         ]
                                                        }
                                            },
                  "/api/dashboard/category-wise":  {
                                                       "get":  {
                                                                   "tags":  [
                                                                                "Dashboard"
                                                                            ],
                                                                   "summary":  "Get category-wise totals",
                                                                   "description":  "Returns total amounts grouped by category. Filter by INCOME or EXPENSE. ANALYST and ADMIN only.",
                                                                   "operationId":  "getCategoryWise",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "type",
                                                                                          "in":  "query",
                                                                                          "required":  false,
                                                                                          "schema":  {
                                                                                                         "type":  "string",
                                                                                                         "default":  "EXPENSE",
                                                                                                         "enum":  [
                                                                                                                      "INCOME",
                                                                                                                      "EXPENSE"
                                                                                                                  ]
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "OK",
                                                                                                 "content":  {
                                                                                                                 "*/*":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/ApiResponseListCategorySummaryResponse"
                                                                                                                                        }
                                                                                                                         }
                                                                                                             }
                                                                                             }
                                                                                 },
                                                                   "security":  [
                                                                                    {
                                                                                        "bearerAuth":  [

                                                                                                       ]
                                                                                    }
                                                                                ]
                                                               }
                                                   },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "ApiResponseListMonthlyTrendResponse":  {
                                                                                   "type":  "object",
                                                                                   "properties":  {
                                                                                                      "success":  {
                                                                                                                      "type":  "boolean"
                                                                                                                  },
                                                                                                      "message":  {
                                                                                                                      "type":  "string"
                                                                                                                  },
                                                                                                      "data":  {
                                                                                                                   "type":  "array",
                                                                                                                   "items":  {
                                                                                                                                 "$ref":  "#/components/schemas/MonthlyTrendResponse"
                                                                                                                             }
                                                                                                               },
                                                                                                      "timestamp":  {
                                                                                                                        "type":  "string",
                                                                                                                        "format":  "date-time"
                                                                                                                    }
                                                                                                  }
                                                                               },
                                       "MonthlyTrendResponse":  {
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "year":  {
                                                                                                    "type":  "integer",
                                                                                                    "format":  "int32"
                                                                                                },
                                                                                       "month":  {
                                                                                                     "type":  "integer",
                                                                                                     "format":  "int32"
                                                                                                 },
                                                                                       "type":  {
                                                                                                    "type":  "string",
                                                                                                    "enum":  [
                                                                                                                 "INCOME",
                                                                                                                 "EXPENSE"
                                                                                                             ]
                                                                                                },
                                                                                       "totalAmount":  {
                                                                                                           "type":  "number"
                                                                                                       }
                                                                                   }
                                                                },
                                       "ApiResponseDashboardSummaryResponse":  {
                                                                                   "type":  "object",
                                                                                   "properties":  {
                                                                                                      "success":  {
                                                                                                                      "type":  "boolean"
                                                                                                                  },
                                                                                                      "message":  {
                                                                                                                      "type":  "string"
                                                                                                                  },
                                                                                                      "data":  {
                                                                                                                   "$ref":  "#/components/schemas/DashboardSummaryResponse"
                                                                                                               },
                                                                                                      "timestamp":  {
                                                                                                                        "type":  "string",
                                                                                                                        "format":  "date-time"
                                                                                                                    }
                                                                                                  }
                                                                               },
                                       "DashboardSummaryResponse":  {
                                                                        "type":  "object",
                                                                        "properties":  {
                                                                                           "totalIncome":  {
                                                                                                               "type":  "number"
                                                                                                           },
                                                                                           "totalExpenses":  {
                                                                                                                 "type":  "number"
                                                                                                             },
                                                                                           "netBalance":  {
                                                                                                              "type":  "number"
                                                                                                          },
                                                                                           "totalTransactions":  {
                                                                                                                     "type":  "integer",
                                                                                                                     "format":  "int64"
                                                                                                                 }
                                                                                       }
                                                                    },
                                       "ApiResponseListTransactionResponse":  {
                                                                                  "type":  "object",
                                                                                  "properties":  {
                                                                                                     "success":  {
                                                                                                                     "type":  "boolean"
                                                                                                                 },
                                                                                                     "message":  {
                                                                                                                     "type":  "string"
                                                                                                                 },
                                                                                                     "data":  {
                                                                                                                  "type":  "array",
                                                                                                                  "items":  {
                                                                                                                                "$ref":  "#/components/schemas/TransactionResponse"
                                                                                                                            }
                                                                                                              },
                                                                                                     "timestamp":  {
                                                                                                                       "type":  "string",
                                                                                                                       "format":  "date-time"
                                                                                                                   }
                                                                                                 }
                                                                              },
                                       "ApiResponseListCategorySummaryResponse":  {
                                                                                      "type":  "object",
                                                                                      "properties":  {
                                                                                                         "success":  {
                                                                                                                         "type":  "boolean"
                                                                                                                     },
                                                                                                         "message":  {
                                                                                                                         "type":  "string"
                                                                                                                     },
                                                                                                         "data":  {
                                                                                                                      "type":  "array",
                                                                                                                      "items":  {
                                                                                                                                    "$ref":  "#/components/schemas/CategorySummaryResponse"
                                                                                                                                }
                                                                                                                  },
                                                                                                         "timestamp":  {
                                                                                                                           "type":  "string",
                                                                                                                           "format":  "date-time"
                                                                                                                       }
                                                                                                     }
                                                                                  },
                                       "CategorySummaryResponse":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "category":  {
                                                                                                           "type":  "string"
                                                                                                       },
                                                                                          "totalAmount":  {
                                                                                                              "type":  "number"
                                                                                                          }
                                                                                      }
                                                                   },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "description":  "JWT Bearer Token. Format: Bearer \u003ctoken\u003e",
                                                                  "in":  "header",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}